+++ Sources/TABResourceLoader/Protocols/NetworkJSONResourceType.swift
@@ -8,9 +8,28 @@
 
 import Foundation
 
+@available(*, deprecated: 2.1.0, message: "Use NetworkJSONDictionaryResourceType or NetworkJSONArrayResourceType")
 public protocol NetworkJSONResourceType: NetworkResourceType, JSONResourceType {}
 
 public extension NetworkJSONResourceType {
+  var HTTPHeaderFields: [String: String]? {
+    return ["Content-Type": "application/json"]
+  }
+}
+
+/// Defines a resource that can be fetched from a network where the root type is a JSON object
+public protocol NetworkJSONDictionaryResourceType: NetworkResourceType, JSONDictionaryResourceType {}
+
+public extension NetworkJSONDictionaryResourceType {
+  var HTTPHeaderFields: [String: String]? {
+    return ["Content-Type": "application/json"]
+  }
+}
+
+/// Defines a resource that can be fetched from a network where the root type is a JSON array
+public protocol NetworkJSONArrayResourceType: NetworkResourceType, JSONArrayResourceType {}
+
+public extension NetworkJSONArrayResourceType {
   var HTTPHeaderFields: [String: String]? {
     return ["Content-Type": "application/json"]
   }